update: more format support for 2epub

+++ transchinese.ts
@@ -1,5 +1,5 @@
 import { basename, dirname } from "jsr:@std/path@^1.0";
-import { getDocument, removeIllegalPath } from "./main.ts";
+import { exists, getDocument, removeIllegalPath, sleep } from "./main.ts";
 import { fetch2 } from "./main.ts";
 import { toEpub } from "./2epub.ts";
 import { ensureDir } from "jsr:@std/fs@^1.0.10/ensure-dir";
@@ -64,7 +64,7 @@ const getNovelByUrl = async (url: URL | string) => {
     return [removeIllegalPath(title), bytes] as [string, Uint8Array];
 }
 
-async function* getIndex (url: URL | string){
+async function* getIndex (url: URL | string): AsyncGenerator<[string, URL]> {
     const doc = await getDocument(url);
     for(const item of doc.querySelectorAll(
         'body > div.md-container > main > div > div.md-content > article > table > tbody > tr > td:nth-child(1) > a'
@@ -79,17 +79,32 @@ async function* getIndex (url: URL | string){
 }
 
 // main
-for await(const [title, url] of getIndex(_url)) try{
-    const [title2, content] = await getNovelByUrl(url);
-    ensureDir(dirname(title2));
-    if(title2.endsWith('.txt'))
-        toEpub(new TextDecoder().decode(content), title2, title2.replace('.txt', '.epub'))
-            .catch((function(this: [string, Uint8Array], e: Error){
-                console.error(e);
-                Deno.writeFile(this[0], this[1]);
-            }).bind([title2, content]));
-    else
-        Deno.writeFile(title2, content);
-}catch(e){
-    console.error(e);
+export async function getTBooks(_url: URL | string){
+    for await(const [title, url] of getIndex(_url)) try{
+        if(await exists(title + '.epub') || await exists(title + '.txt')){
+            console.log(title, 'already exists, skipping...');
+            continue;
+        } 
+        try{
+            // deno-lint-ignore no-inner-declarations
+            var __res = await getNovelByUrl(url);
+        }catch{
+            console.log(title, 'seemed to be a directory, recursing...');
+            await getTBooks(url);
+            continue;
+        }
+        const [title2, content] = __res;
+        ensureDir(dirname(title2));
+        if(title2.endsWith('.txt') && toEpub(new TextDecoder().decode(content), title2, title2.replace('.txt', '.epub'))){
+            // success
+        }else{
+            Deno.writeFile(title2, content);
+        }
+
+        // await sleep(1); // wait downloadings
+    }catch(e){
+        console.error(e);
+    }
 }
+
+if(import.meta.main) getTBooks(_url);
